Change the flow of DOM.
In order to accomodate the passing of names to gamecontroller. I added some functions in it.
I also changed the flow of DOM play. Now it will start with input of names, then form listener, then game start. So the values are checked first before being used in the game.